Move actions to separate files and colocate tests
