+++ .mocharc.js
@@ -1,7 +1,4 @@
 module.exports = {
-	file: [
-		'test/test.ts'
-	],
 	require: [
 		'sucrase/register'
 	]
