+++ package.json
@@ -9,7 +9,7 @@
     "preversion": "npm run build",
     "version": "auto-changelog -p --template keepachangelog && git add CHANGELOG.md",
     "prepublishOnly": "git push && git push --tags && gh-release",
-    "test": "mocha"
+    "test": "mocha ./src/test.ts"
   },
   "keywords": [
     "svelte"
