+++ src/test.ts
@@ -4,4 +4,4 @@ const window = new jsdom.JSDOM('<main></main>').window;
 global.document = window.document;
 global.window = window;
 
-require('./clickOutside');
+require('./clickOutside.test');
