+++ src/types.ts
@@ -1,0 +1,4 @@
+export type Action = (node: HTMLElement, parameters: any) => {
+	update?: (parameters: any) => void,
+	destroy?: () => void
+}
